rename FilteredEventsMap to EventFilters
