+++ src/app/events/page.tsx
@@ -4,7 +4,7 @@ import type { NextPage } from "next";
 
 import { Title } from "@/components/shared/Title";
 import { Slider } from "@/components/shared/Slider";
-import { FilteredEventsMap } from "@/components/widgets/FilteredEventsMap";
+import { EventFilters } from "@/components/widgets/EventFilters";
 
 import { CreateAndInvite } from "@/components/widgets/CreateAndInvite";
 import { useSession } from "next-auth/react";
@@ -21,7 +21,7 @@ const EventListPage: NextPage = () => {
         <div className={classes.eventListPage}>
             <Title tag={"h1"} title="Join the adventure" className={classes.title} />
             <Slider />
-            <FilteredEventsMap />
+            <EventFilters />
             <CreateAndInvite mode={Mode.EVENT} status={status} />
         </div>
     );
