+++ src/components/widgets/EventFilters/FilterEvents/FilterEvents.module.css
