+++ src/components/widgets/EventFilters/FilterEvents/FilterEvents.tsx
