+++ src/components/widgets/EventFilters/FilterEvents/index.ts
