+++ src/components/widgets/EventFilters/FilteredEventsMap.module.css
