+++ src/components/widgets/EventFilters/NavbarEvents/NavbarEvents.tsx
