+++ src/components/widgets/EventFilters/index.ts
@@ -1,0 +1,1 @@
+export { EventFilters } from "./EventFilters";
