--- src/components/widgets/FilteredEventsMap/index.ts
@@ -1,1 +1,0 @@
-export { FilteredEventsMap } from "./FilteredEventsMap";
